backend/requirements.txt
fastapi
pydantic
google-generativeai
python-dotenv
uvicorn[standard]
openai